Day 5 plus One to vector number | 100% Beats |
https://leetcode.com/problems/plus-one/solutions/4193589/cpp-simple-solution-recursion-runtime-0sec-100-beats/
please visit for more information.

The approach used in this code is to recursively add 1 to the digits in the digits vector, starting from the rightmost digit. If the rightmost digit is less than 9, then 1 is simply added to it. If the rightmost digit is 9, then it is set to 0, and 1 is added to the next digit to the left. This process is repeated until the leftmost digit is reached. If the leftmost digit is also 9, then a new digit of 1 is added to the front of the vector.