94. Bianry Tree Inorder Traversal
https://leetcode.com/problems/binary-tree-inorder-traversal/solutions/4196973/cpp-easy-solution-recursion-100-beats/

The algorithm starts at the root node of the binary tree. It then recursively visits the left subtree of the root node. As each node in the left subtree is visited, its value is added to the vector. Once the left subtree has been visited, the algorithm visits the root node and adds its value to the vector. The algorithm then recursively visits the right subtree of the root node. As each node in the right subtree is visited, its value is added to the vector. Once the right subtree has been visited, the algorithm returns the vector.